move messaging-event to standalone Django view
