+++ corehq/apps/api/resources/v0_5.py
@@ -1093,366 +1093,3 @@ class ODataFormResource(BaseODataResource):
             url(r"^(?P<resource_name>{})/(?P<config_id>[\w\d_.-]+)/feed".format(
                 self._meta.resource_name), self.wrap_view('dispatch_list')),
         ]
-
-
-class MessagingEventResource(HqBaseResource, ModelResource):
-    """Despite it's name this API is backed by the MessagingSubEvent model
-    which has a more direct relationship with who the messages are being sent to.
-    Each event may have more than one actual message associated with it.
-    """
-
-    source = fields.DictField()
-    recipient = fields.DictField()
-    form = fields.DictField()
-    error = fields.DictField()
-    messages = fields.ListField()
-
-    def dehydrate(self, bundle):
-        bundle.data["domain"] = bundle.obj.parent.domain
-        return bundle
-
-    def dehydrate_status(self, bundle):
-        """The status of the event is not tied to the status of individual messages. An event
-        may be 'complete' but some or all of the messages may have errored.
-
-        In the case of completed survey events the status is taken from the XFormsSession.
-        """
-        event = bundle.obj
-        if event.status == MessagingEvent.STATUS_COMPLETED and event.xforms_session_id:
-            return event.xforms_session.status_slug
-        return MessagingEvent.STATUS_SLUGS.get(event.status, 'unknown')
-
-    def dehydrate_content_type(self, bundle):
-        return MessagingEvent.CONTENT_TYPE_SLUGS.get(bundle.obj.content_type, "unknown")
-
-    def dehydrate_source(self, bundle):
-        """This is the 'trigger' for the event e.g. broadcast, conditional-alert etc."""
-        parent = bundle.obj.parent
-
-        return {
-            "id": parent.source_id,
-            "type": MessagingEvent.SOURCE_SLUGS.get(parent.source, 'unknown'),
-            "name": get_event_display_api(parent),
-        }
-
-    def dehydrate_recipient(self, bundle):
-        display_value = None
-        if bundle.obj.recipient_id:
-            display_value = case_or_user_id_to_name(bundle.obj.recipient_id, {
-                "couch_recipient_doc_type": bundle.obj.get_recipient_doc_type()
-            })
-        return {
-            "id": bundle.obj.recipient_id,
-            "type": MessagingSubEvent.RECIPIENT_SLUGS.get(bundle.obj.recipient_type, "unknown"),
-            "name": display_value or "unknown",
-        }
-
-    def dehydrate_form(self, bundle):
-        event = bundle.obj
-        if event.content_type not in (MessagingEvent.CONTENT_SMS_SURVEY, MessagingEvent.CONTENT_IVR_SURVEY):
-            return None
-
-        submission_id = None
-        if event.xforms_session_id:
-            submission_id = event.xforms_session.submission_id
-        return {
-            "app_id": bundle.obj.app_id,
-            "form_definition_id": bundle.obj.form_unique_id,
-            "form_name": bundle.obj.form_name,
-            "form_submission_id": submission_id,
-        }
-
-    def dehydrate_error(self, bundle):
-        """Details about the error at the event level. Not related to errors of individual messages."""
-        event = bundle.obj
-        if not event.error_code:
-            return None
-
-        return {
-            "code": event.error_code,
-            "message": MessagingEvent.ERROR_MESSAGES.get(event.error_code, None),
-            "message_detail": event.additional_error_text
-        }
-
-    def dehydrate_messages(self, bundle):
-        """The content_types supported here are the only ones used by the MessagingSubEvent
-        model. Other content types such as CONTENT_CHAT_SMS are used by the MessagingEvent model.
-        """
-        event = bundle.obj
-        if event.content_type == MessagingEvent.CONTENT_EMAIL:
-            return self._get_messages_for_email(event)
-
-        if event.content_type in (MessagingEvent.CONTENT_SMS, MessagingEvent.CONTENT_SMS_CALLBACK):
-            return self._get_messages_for_sms(event)
-
-        if event.content_type in (MessagingEvent.CONTENT_SMS_SURVEY, MessagingEvent.CONTENT_IVR_SURVEY):
-            return self._get_messages_for_survey(event)
-        return []
-
-    def _get_messages_for_email(self, event):
-        try:
-            email = Email.objects.get(messaging_subevent=event.pk)
-            content = email.body
-            recipient_address = email.recipient_address
-        except Email.DoesNotExist:
-            content = '-'
-            recipient_address = '-'
-
-        return [{
-            "date": event.date,
-            "type": "email",
-            "direction": "outgoing",
-            "content": content,
-            "status": MessagingEvent.STATUS_SLUGS.get(event.status, 'unknown'),
-            "backend": "email",
-            "contact": recipient_address
-        }]
-
-    def _get_messages_for_sms(self, event):
-        messages = SMS.objects.filter(messaging_subevent_id=event.pk)
-        return self._get_message_dicts_for_sms(event, messages, "sms")
-
-    def _get_messages_for_survey(self, event):
-        if not event.xforms_session_id:
-            return []
-
-        xforms_session = event.xforms_session
-        if not xforms_session:
-            return []
-
-        messages = SMS.objects.filter(xforms_session_couch_id=xforms_session.couch_id)
-        type_ = "ivr" if event.content_type == MessagingEvent.CONTENT_IVR_SURVEY else "sms"
-        return self._get_message_dicts_for_sms(event, messages, type_)
-
-    def _get_message_dicts_for_sms(self, event, messages, type_):
-        message_dicts = []
-        for sms in messages:
-            error_message = None
-            if event.status != MessagingEvent.STATUS_ERROR:
-                status, error_message = get_sms_status_display_raw(sms)
-            else:
-                status = MessagingEvent.STATUS_SLUGS.get(event.status, "unknown")
-
-            message_data = {
-                "date": sms.date,
-                "type": type_,
-                "direction": SMS.DIRECTION_SLUGS.get(sms.direction, "unknown"),
-                "content": sms.text,
-                "status": status,
-                "backend": get_backend_name(sms.backend_id) or sms.backend_id,
-                "contact": sms.phone_number
-            }
-            if error_message:
-                message_data["error_message"] = error_message
-            message_dicts.append(message_data)
-        return message_dicts
-
-    def build_filters(self, filters=None, **kwargs):
-        """Support custom filtering. Consumers must return None or dictionary with
-        Django filter kwargs or else a Django Q object:
-
-        return {"date__lte": value}
-        return {"date": Q(date__lte=value) | Q(date__gt=value)}
-        """
-        filter_consumers = [
-            self._get_date_filter_consumer(),
-            self._get_source_filter_consumer(),
-            self._get_content_type_filter_consumer(),
-            self._status_filter_consumer,
-            self._get_error_code_filter_consumer(),
-            self._get_email_filter_consumer(),
-            self._get_phone_number_filter_consumer(),
-        ]
-        orm_filters = {}
-        for key, value in list(filters.items()):
-            for consumer in filter_consumers:
-                result = consumer(key, value)
-                if result:
-                    del filters[key]
-                    orm_filters.update(result)
-                    continue
-        orm_filters.update(super().build_filters(filters, **kwargs))
-        return orm_filters
-
-    def apply_filters(self, request, applicable_filters):
-        """Separate out Django kwargs filters from Q filters and apply them
-        to the query."""
-        native_filters = []
-        for key, value in list(applicable_filters.items()):
-            if isinstance(value, Q):
-                native_filters.append(applicable_filters.pop(key))
-        query = self.get_object_list(request).filter(**applicable_filters)
-        if native_filters:
-            query = query.filter(*native_filters)
-        return query
-
-    @staticmethod
-    def _get_date_filter_consumer():
-        """date.{lt, lte, gt, gte}=<ISO DATE>"""
-        date_filter = make_date_filter(functools.partial(django_date_filter, field_name="date"))
-
-        def _date_consumer(key, value):
-            if '.' in key and key.split(".")[0] == "date":
-                prefix, qualifier = key.split(".", maxsplit=1)
-                try:
-                    return date_filter(qualifier, value)
-                except ValueError as e:
-                    raise InvalidFilterError(str(e))
-
-        return _date_consumer
-
-    @staticmethod
-    def _get_source_filter_consumer():
-        """source=<SOURCE SLUG>"""
-        # match functionality in corehq.apps.reports.standard.sms.MessagingEventsReport.get_filters
-        expansions = {
-            MessagingEvent.SOURCE_OTHER: [MessagingEvent.SOURCE_FORWARDED],
-            MessagingEvent.SOURCE_BROADCAST: [
-                MessagingEvent.SOURCE_SCHEDULED_BROADCAST,
-                MessagingEvent.SOURCE_IMMEDIATE_BROADCAST
-            ],
-            MessagingEvent.SOURCE_REMINDER: [MessagingEvent.SOURCE_CASE_RULE]
-        }
-        return MessagingEventResource._make_slug_filter_consumer(
-            "source", MessagingEvent.SOURCE_SLUGS, "parent__source__in", expansions
-        )
-
-    @staticmethod
-    def _get_content_type_filter_consumer():
-        """content_type=<CONTENT TYPE SLUG>"""
-        # match functionality in corehq.apps.reports.standard.sms.MessagingEventsReport.get_filters
-        expansions = {
-            MessagingEvent.CONTENT_SMS_SURVEY: [
-                MessagingEvent.CONTENT_SMS_SURVEY,
-                MessagingEvent.CONTENT_IVR_SURVEY,
-            ],
-            MessagingEvent.CONTENT_SMS: [
-                MessagingEvent.CONTENT_SMS,
-                MessagingEvent.CONTENT_PHONE_VERIFICATION,
-                MessagingEvent.CONTENT_ADHOC_SMS,
-                MessagingEvent.CONTENT_API_SMS,
-                MessagingEvent.CONTENT_CHAT_SMS
-            ],
-        }
-        return MessagingEventResource._make_slug_filter_consumer(
-            "content_type", MessagingEvent.CONTENT_TYPE_SLUGS, "content_type__in", expansions
-        )
-
-    @staticmethod
-    def _make_slug_filter_consumer(filter_key, slug_dict, model_filter_arg, expansions=None):
-        slug_values = {v: k for k, v in slug_dict.items()}
-
-        def _consumer(key, value):
-            if key != filter_key:
-                return
-
-            values = value.split(',')
-            vals = [slug_values[val] for val in values if val in slug_values]
-            if vals:
-                for key, extras in (expansions or {}).items():
-                    if key in vals:
-                        vals.extend(extras)
-                return {model_filter_arg: vals}
-
-        return _consumer
-
-    def _status_filter_consumer(self, key, value):
-        """status=<STATUS SLUG>
-
-        Status filtering is applied to the event as well as to the messages / XFormSession.
-        """
-        slug_values = {v: k for k, v in MessagingEvent.STATUS_SLUGS.items()}
-        if key != "status":
-            return
-
-        model_value = slug_values.get(value, value)
-        # match functionality in corehq.pps.reports.standard.sms.MessagingEventsReport.get_filters
-        if model_value == MessagingEvent.STATUS_ERROR:
-            return {"status": (Q(status=model_value) | Q(sms__error=True))}
-        elif model_value == MessagingEvent.STATUS_IN_PROGRESS:
-            # We need to check for id__isnull=False below because the
-            # query we make in this report has to do a left join, and
-            # in this particular filter we can only validly check
-            # session_is_open=True if there actually are
-            # subevent and xforms session records
-            return {"status": (
-                Q(status=model_value)
-                | (Q(xforms_session__id__isnull=False) & Q(xforms_session__session_is_open=True))
-            )}
-        elif model_value == MessagingEvent.STATUS_NOT_COMPLETED:
-            return {"status": (
-                Q(status=model_value)
-                | (Q(xforms_session__session_is_open=False) & Q(xforms_session__submission_id__isnull=True))
-            )}
-        elif model_value == MessagingEvent.STATUS_EMAIL_DELIVERED:
-            return {"status": model_value}
-        else:
-            raise InvalidFilterError(f"'{value}' is an invalid value for the 'status' filter")
-
-    def _get_error_code_filter_consumer(self):
-        """error_code=<ERROR CODE>
-
-        This has to be done here since `error_code` is not a field in the API so Tastypie
-        won't allow filtering against it.
-        """
-        return self._make_simple_consumer("error_code", "error_code")
-
-    def _get_email_filter_consumer(self):
-        """email_address=<EMAIL>
-
-        Values are validated prior to being applied.
-        """
-        return self._make_simple_consumer(
-            "email_address", "email__recipient_address", validate_email
-        )
-
-    def _get_phone_number_filter_consumer(self):
-        """phone_number=<PHONE NUMBER>"""
-        return self._make_simple_consumer(
-            "phone_number", "sms__phone_number__contains", validate_phone_number
-        )
-
-    @staticmethod
-    def _make_simple_consumer(filter_key, model_filter_arg, validator=None):
-        def _consumer(key, value):
-            if key != filter_key:
-                return
-
-            value = value.strip()
-
-            if validator:
-                try:
-                    validator(value)
-                except ValidationError:
-                    raise InvalidFilterError(f"'{value}' is invalid value for the '{filter_key}' filter")
-
-            return {model_filter_arg: value}
-        return _consumer
-
-    class Meta(object):
-        queryset = MessagingSubEvent.objects.select_related("parent").all()
-        include_resource_uri = False
-        list_allowed_methods = ['get']
-        detail_allowed_methods = ['get']
-        resource_name = 'messaging-event'
-        authentication = RequirePermissionAuthentication(Permissions.edit_data)
-        authorization = DomainAuthorization('parent__domain')
-        paginator_class = NoCountingPaginator
-        serializer = Serializer(formats=['json'])
-        excludes = {
-            "error_code",
-            "additional_error_text",
-            "app_id",
-            "form_name",
-            "form_unique_id",
-            "recipient_id",
-            "recipient_type",
-        }
-        filtering = {
-            # this is needed for the domain filtering but any values passed in via the URL get overridden
-            "domain": ('exact',),
-            "case_id": ('exact',),
-        }
-        ordering = [
-            'date',
-        ]
